refacor: asset service interfaces
