+++ server/libs/common/src/services/asset.service.ts
@@ -1,0 +1,15 @@
+import { AssetType } from '../models';
+
+export type ServerUsageItems = ServerUsageItem[];
+export interface ServerUsageItem {
+  userId: string;
+  assetType: AssetType;
+  assetCount: number;
+  totalSizeInBytes: number;
+}
+
+export const IAssetRepository = 'AssetRepository';
+
+export interface IAssetRepository {
+  getUserStats(): Promise<ServerUsageItems>;
+}
